requirements-dev.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
django-cors-headers==3.7.0
django-markdownx==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.26.5
